[PINT-1145] Update default attribute options (#31)
The product attribute used to determine if products should be eligible for Fast checkout appears on the front-end if not manually removed from listing in the Magento admin panel. We should make it not be listed in Layered Navigation by default so Sellers don’t have to change this setting to avoid Fast being listed.